+++ tests/test_response_processor.py
@@ -164,3 +164,28 @@ class TestResponseProcessor(unittest.TestCase):
 
         with self.assertRaises(CTPQueue):
             self.rp.send_receipt(valid_json)
+
+    def test_service_name_return_responses(self):
+        url = "www.testing.test/responses"
+        service = self.rp.service_name(url)
+        self.assertEqual(service, 'SDX-STORE')
+
+    def test_service_name_return_decrypt(self):
+        url = "www.testing.test/decrypt"
+        service = self.rp.service_name(url)
+        self.assertEqual(service, 'SDX-DECRYPT')
+
+    def test_service_name_return_validate(self):
+        url = "www.testing.test/validate"
+        service = self.rp.service_name(url)
+        self.assertEqual(service, 'SDX-VALIDATE')
+
+    def test_service_name_return_none(self):
+        url = "www.testing.test/test/12345"
+        service = self.rp.service_name(url)
+        self.assertEqual(service, None)
+
+    def test_url_service_name_none(self):
+        url = None
+        service = self.rp.service_name(url)
+        self.assertEqual(service, None)